+++ CHANGELOG.md
@@ -11,6 +11,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `<slot name="toolbar">` to `<leaflet-map>`
 - `<form is="share-target">` for handling shares using the Share Trarget API
 
+### Fixed
+- Properly handle parsing share target params [#207](https://github.com/shgysk8zer0/cdn.kernvalley.us/issues/207)
+
 ## [v1.0.7] - 2020-07-20
 
 ### Fixed